Added user permissions

+++ app/Http/Livewire/Backoffice/Customers/Delete.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Livewire\Backoffice\Customers;
 
-use App\Models\Customer;
 use Livewire\Component;
+use App\Models\Customer;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
 
     public $customer;
 
@@ -17,6 +19,7 @@ class Delete extends Component
 
     public function delete()
     {
+        $this->authorize('delete', $this->customer);
         $this->customer->delete();
         $this->emit('success', __('customers.success_delete'), sprintf(__('customers.message_delete'), $this->customer->business_name));
         $this->emit('hideModal');
