+++ app/Http/Livewire/Backoffice/Users/Delete.php
@@ -4,9 +4,12 @@ namespace App\Http\Livewire\Backoffice\Users;
 
 use App\Models\User;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
+
     public $user;
 
     public function mount(User $user)
@@ -16,6 +19,7 @@ class Delete extends Component
 
     public function delete()
     {
+        $this->authorize('delete', $this->user);
         $this->user->delete();
         $this->emit('success', __('users.success_delete'), sprintf(__('users.message_delete'), $this->user->email));
         $this->emit('hideModal');
