+++ app/Http/Livewire/Backoffice/Zones/Delete.php
@@ -4,9 +4,11 @@ namespace App\Http\Livewire\Backoffice\Zones;
 
 use App\Models\Zone;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
 
     public $zone;
 
@@ -17,6 +19,7 @@ class Delete extends Component
 
     public function delete()
     {
+        $this->authorize('delete', $this->zone);
         $this->zone->delete();
         $this->emit('success', __('zones.zones.alert.delete.success'), sprintf(__('zones.zones.alert.delete.message'), $this->zone->name));
         $this->emit('hideModal');
